+++ tests/testthat/test-calc_distance.R
@@ -1,4 +1,9 @@
 test_that("calc_distance works", {
-  expect_type(calc_distance(evavelo_example),
-                  "list")
+
+  expect_type(calc_distance(evavelo_example_geocoded),
+                  "list") %>%
+    suppressMessages()
+
+  expect_error(calc_distance(evavelo_example),
+               "Cannot calculate distance on evadata that is not geocoded")
 })